Refactor web dialogs launcher sharing
