+++ filekit-dialogs-compose/src/jvmAndNativeMain/kotlin/io/github/vinceglb/filekit/dialogs/compose/FileKitCompose.nativeAndJvm.kt
@@ -8,56 +8,23 @@ import androidx.compose.runtime.rememberUpdatedState
 import io.github.vinceglb.filekit.FileKit
 import io.github.vinceglb.filekit.PlatformFile
 import io.github.vinceglb.filekit.dialogs.FileKitDialogSettings
-import io.github.vinceglb.filekit.dialogs.openDirectoryPicker
 import io.github.vinceglb.filekit.dialogs.openFileSaver
 import kotlinx.coroutines.launch
 
 /**
- * Creates and remembers a [PickerResultLauncher] for picking a directory.
+ * Creates a launcher for the file saver dialog.
  *
- * @param directory The initial directory. Supported on desktop platforms.
  * @param dialogSettings Platform-specific settings for the dialog.
- * @param onResult Callback invoked with the picked directory, or null if cancelled.
- * @return A [PickerResultLauncher] that can be used to launch the picker.
+ * @param onResult Callback to receive the result of the file saver dialog.
+ * @return A [SaverResultLauncher] instance.
  */
-@Composable
-public actual fun rememberDirectoryPickerLauncher(
-    directory: PlatformFile?,
-    dialogSettings: FileKitDialogSettings,
-    onResult: (PlatformFile?) -> Unit,
-): PickerResultLauncher {
-    // Init FileKit
-    InitFileKit()
-
-    // Coroutine
-    val coroutineScope = rememberCoroutineScope()
-
-    // Updated state
-    val currentDirectory by rememberUpdatedState(directory)
-    val currentOnResult by rememberUpdatedState(onResult)
-
-    // FileKit launcher
-    val returnedLauncher = remember {
-        PickerResultLauncher {
-            coroutineScope.launch {
-                val result = FileKit.openDirectoryPicker(
-                    directory = currentDirectory,
-                    dialogSettings = dialogSettings,
-                )
-                currentOnResult(result)
-            }
-        }
-    }
-
-    return returnedLauncher
-}
-
 @Composable
 internal actual fun rememberPlatformFileSaverLauncher(
     dialogSettings: FileKitDialogSettings,
     onResult: (PlatformFile?) -> Unit,
 ): SaverResultLauncher {
     val coroutineScope = rememberCoroutineScope()
+    val currentDialogSettings by rememberUpdatedState(dialogSettings)
     val currentOnResult by rememberUpdatedState(onResult)
 
     return remember {
@@ -67,7 +34,7 @@ internal actual fun rememberPlatformFileSaverLauncher(
                     suggestedName = suggestedName,
                     extension = extension,
                     directory = directory,
-                    dialogSettings = dialogSettings,
+                    dialogSettings = currentDialogSettings,
                 )
                 currentOnResult(result)
             }
