+++ filekit-dialogs-compose/src/nonAndroidMain/kotlin/io/github/vinceglb/filekit/dialogs/compose/FileKitCompose.nonAndroid.kt
@@ -10,8 +10,44 @@ import io.github.vinceglb.filekit.PlatformFile
 import io.github.vinceglb.filekit.dialogs.FileKitDialogSettings
 import io.github.vinceglb.filekit.dialogs.FileKitMode
 import io.github.vinceglb.filekit.dialogs.FileKitType
+import io.github.vinceglb.filekit.dialogs.openDirectoryPicker
 import io.github.vinceglb.filekit.dialogs.openFilePicker
 import kotlinx.coroutines.launch
+
+/**
+ * Creates and remembers a [PickerResultLauncher] for picking a directory.
+ *
+ * @param directory The initial directory. Supported on desktop platforms.
+ * @param dialogSettings Platform-specific settings for the dialog.
+ * @param onResult Callback invoked with the picked directory, or null if cancelled.
+ * @return A [PickerResultLauncher] that can be used to launch the picker.
+ */
+@Composable
+public actual fun rememberDirectoryPickerLauncher(
+    directory: PlatformFile?,
+    dialogSettings: FileKitDialogSettings,
+    onResult: (PlatformFile?) -> Unit,
+): PickerResultLauncher {
+    InitFileKit()
+
+    val coroutineScope = rememberCoroutineScope()
+
+    val currentDirectory by rememberUpdatedState(directory)
+    val currentDialogSettings by rememberUpdatedState(dialogSettings)
+    val currentOnResult by rememberUpdatedState(onResult)
+
+    return remember {
+        PickerResultLauncher {
+            coroutineScope.launch {
+                val result = FileKit.openDirectoryPicker(
+                    directory = currentDirectory,
+                    dialogSettings = currentDialogSettings,
+                )
+                currentOnResult(result)
+            }
+        }
+    }
+}
 
 @Composable
 internal actual fun <PickerResult, ConsumedResult> rememberPlatformFilePickerLauncher(
@@ -26,6 +62,7 @@ internal actual fun <PickerResult, ConsumedResult> rememberPlatformFilePickerLau
     val currentType by rememberUpdatedState(type)
     val currentMode by rememberUpdatedState(mode)
     val currentDirectory by rememberUpdatedState(directory)
+    val currentDialogSettings by rememberUpdatedState(dialogSettings)
     val currentOnConsumed by rememberUpdatedState(onResult)
 
     return remember {
@@ -35,7 +72,7 @@ internal actual fun <PickerResult, ConsumedResult> rememberPlatformFilePickerLau
                     type = currentType,
                     mode = currentMode,
                     directory = currentDirectory,
-                    dialogSettings = dialogSettings,
+                    dialogSettings = currentDialogSettings,
                 )
                 currentMode.consumeResult(result, currentOnConsumed)
             }
