+++ filekit-dialogs-compose/src/webMain/kotlin/io/github/vinceglb/filekit/dialogs/compose/FileKitCompose.web.kt
@@ -1,55 +1,6 @@
 package io.github.vinceglb.filekit.dialogs.compose
 
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.rememberUpdatedState
-import io.github.vinceglb.filekit.FileKit
-import io.github.vinceglb.filekit.PlatformFile
-import io.github.vinceglb.filekit.dialogs.FileKitDialogSettings
-import io.github.vinceglb.filekit.dialogs.openDirectoryPicker
-import kotlinx.coroutines.launch
 
 @Composable
 internal actual fun InitFileKit() {}
-
-/**
- * Creates and remembers a [PickerResultLauncher] for picking a directory.
- *
- * @param directory The initial directory. Supported on desktop platforms.
- * @param dialogSettings Platform-specific settings for the dialog.
- * @param onResult Callback invoked with the picked directory, or null if cancelled.
- * @return A [PickerResultLauncher] that can be used to launch the picker.
- */
-@Composable
-public actual fun rememberDirectoryPickerLauncher(
-    directory: PlatformFile?,
-    dialogSettings: FileKitDialogSettings,
-    onResult: (PlatformFile?) -> Unit,
-): PickerResultLauncher {
-    // Init FileKit
-    InitFileKit()
-
-    // Coroutine
-    val coroutineScope = rememberCoroutineScope()
-
-    // Updated state
-    val currentDirectory by rememberUpdatedState(directory)
-    val currentOnResult by rememberUpdatedState(onResult)
-
-    // FileKit launcher
-    val returnedLauncher = remember {
-        PickerResultLauncher {
-            coroutineScope.launch {
-                val result = FileKit.openDirectoryPicker(
-                    directory = currentDirectory,
-                    dialogSettings = dialogSettings,
-                )
-                currentOnResult(result)
-            }
-        }
-    }
-
-    return returnedLauncher
-}
