+++ filekit-dialogs/src/webMain/kotlin/io/github/vinceglb/filekit/dialogs/FileKitDialogSettings.web.kt
@@ -1,8 +1,8 @@
 package io.github.vinceglb.filekit.dialogs
 
 /**
- * WasmJs implementation of [FileKitDialogSettings].
- * Currently, there are no specific settings for WasmJs file dialogs.
+ * Web implementation of [FileKitDialogSettings].
+ * Currently, there are no specific settings for web file dialogs.
  */
 public actual class FileKitDialogSettings {
     public actual companion object {
